Add ReferenceTable logic, fetch when reference added
