+++ frontend/src/components/ReferenceTable/ReferenceTableHead.js
@@ -1,0 +1,110 @@
+import * as React from 'react'
+import PropTypes from 'prop-types'
+import Box from '@mui/material/Box'
+import TableCell from '@mui/material/TableCell'
+import TableRow from '@mui/material/TableRow'
+import Checkbox from '@mui/material/Checkbox'
+import TableHead from '@mui/material/TableHead'
+import TableSortLabel from '@mui/material/TableSortLabel'
+import { visuallyHidden } from '@mui/utils'
+
+const headCells = [
+    {
+        id: 'authors',
+        numeric: false,
+        disablePadding: true,
+        label: 'Authors',
+    },
+    {
+        id: 'title',
+        numeric: true,
+        disablePadding: false,
+        label: 'Title',
+    },
+    {
+        id: 'year',
+        numeric: true,
+        disablePadding: false,
+        label: 'Year',
+    },
+    {
+        id: 'publisher',
+        numeric: true,
+        disablePadding: false,
+        label: 'Publisher',
+    },
+    {
+        id: 'publisherAddress',
+        numeric: true,
+        disablePadding: false,
+        label: 'Publisher Address',
+    },
+]
+
+const ReferenceTableHead = (props) => {
+    const {
+        onSelectAllClick,
+        order,
+        orderBy,
+        numSelected,
+        rowCount,
+        onRequestSort,
+    } = props
+    const createSortHandler = (property) => (event) => {
+        onRequestSort(event, property)
+    }
+
+    return (
+        <TableHead>
+            <TableRow>
+                <TableCell padding="checkbox">
+                    <Checkbox
+                        color="primary"
+                        indeterminate={
+                            numSelected > 0 && numSelected < rowCount
+                        }
+                        checked={rowCount > 0 && numSelected === rowCount}
+                        onChange={onSelectAllClick}
+                        inputProps={{
+                            'aria-label': 'select all references',
+                        }}
+                    />
+                </TableCell>
+                {headCells.map((headCell) => (
+                    <TableCell
+                        key={headCell.id}
+                        align={headCell.numeric ? 'right' : 'left'}
+                        padding={headCell.disablePadding ? 'none' : 'normal'}
+                        sortDirection={orderBy === headCell.id ? order : false}
+                    >
+                        <TableSortLabel
+                            active={orderBy === headCell.id}
+                            direction={orderBy === headCell.id ? order : 'asc'}
+                            onClick={createSortHandler(headCell.id)}
+                        >
+                            {headCell.label}
+                            {orderBy === headCell.id ? (
+                                <Box component="span" sx={visuallyHidden}>
+                                    {order === 'desc'
+                                        ? 'sorted descending'
+                                        : 'sorted ascending'}
+                                </Box>
+                            ) : null}
+                        </TableSortLabel>
+                    </TableCell>
+                ))}
+            </TableRow>
+        </TableHead>
+    )
+}
+
+ReferenceTableHead.propTypes = {
+    numSelected: PropTypes.number.isRequired,
+    onRequestSort: PropTypes.func.isRequired,
+    onSelectAllClick: PropTypes.func.isRequired,
+    order: PropTypes.oneOf(['asc', 'desc']).isRequired,
+    orderBy: PropTypes.string.isRequired,
+    rowCount: PropTypes.number.isRequired,
+}
+
+export default ReferenceTableHead
