+++ src/routes.py
@@ -3,30 +3,25 @@ from flask import jsonify, request, send_from_directory
 from services.reference_services import reference_service
 import dataprocessing # REFACTOR: routes.py should only use the services class
 
-
 @app.route("/")
 def index():
     return app.send_static_file("index.html")
 
-# adding reference
-
-
 @app.route("/add_reference", methods=["POST"])
 def add_reference():
     if request.method == "POST":
-        author = request.form["author"]
+        authors = request.form["authors"]
         title = request.form["title"]
         year = request.form["year"]
         publisher = request.form["publisher"]
         publisher_address = request.form["publisher_address"]
 
-        print(author, title, year, publisher, publisher_address)
-
-        if dataprocessing.add_book(author, title, year,
+        if dataprocessing.add_book(authors, title, year,
                                    publisher, publisher_address):
-            return jsonify({"message": "Reference added"})
+            return jsonify({"message": "Reference added"}), 201
         else:
-            return jsonify({"message": "Error occurred when adding reference"})
+            return jsonify({
+                "message": "Error occurred when adding reference"}), 500
 
 @app.route("/get_all_references")
 def get_all_references():
@@ -35,7 +30,6 @@ def get_all_references():
 @app.route('/<path:path>')
 def serve_static(path):
     return send_from_directory('build', path)
-
 
 if __name__ == "__main__":
     app.run(debug=True)
